liver-segmentation-mvp/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
onnxruntime
numpy
pydicom
SimpleITK
Pillow
pyvista
opencv-python-headless


